eliminated stupid compiler warnings

+++ mac/Classes/Editors/TextMateEditor.m
@@ -18,7 +18,7 @@
 - (BOOL)jumpToFile:(NSString *)file line:(NSInteger)line {
     NSString *url;
     if (line > 0)
-        url = [NSString stringWithFormat:@"txmt://open/?url=file://%@&line=%d", [file stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding], line];
+        url = [NSString stringWithFormat:@"txmt://open/?url=file://%@&line=%ld", [file stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding], (long)line];
     else
         url = [NSString stringWithFormat:@"txmt://open/?url=file://%@", [file stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding]];
     [[NSWorkspace sharedWorkspace] openURL:[NSURL URLWithString:url]];
