+++ mac/Classes/Monitoring/OldFSMonitor.m
@@ -197,7 +197,7 @@ static void FSMonitorEventStreamCallback(ConstFSEventStreamRef streamRef, FSMoni
 
 
 static void FSMonitorEventStreamCallback(ConstFSEventStreamRef streamRef, FSMonitor *monitor, size_t numEvents, NSArray *eventPaths, const FSEventStreamEventFlags eventFlags[], const FSEventStreamEventId eventIds[]) {
-    for (int i = 0; i < numEvents; i++) {
+    for (int i = 0; i < (int)numEvents; i++) {
         [monitor sendChangeEventWithPath:[eventPaths objectAtIndex:i] flags:eventFlags[i]];
     }
 }
